requirements.txt
absl-py==2.2.2
altair==4.2.2
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.0.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
entrypoints==0.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grad-cam==1.5.5
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
narwhals==1.37.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==3.19.6
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
pydicom==3.0.1
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
six==1.17.0
smmap==5.0.2
streamlit==1.9.2
sympy==1.13.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
ttach==0.0.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
